Replace datatype for cpu usage with shorts
Also fixed a rendering bug

+++ src/main/java/taskmanager/InformationLoader.java
@@ -33,7 +33,7 @@ public abstract class InformationLoader {
 
 		systemInformation.cpuUsagePerCore = new MeasurementContainer[systemInformation.logicalProcessorCount];
 		for (int i = 0; i < systemInformation.cpuUsagePerCore.length; i++) {
-			systemInformation.cpuUsagePerCore[i] = new MeasurementContainer<>(0L);
+			systemInformation.cpuUsagePerCore[i] = new MeasurementContainer<>((short) 0);
 		}
 
 		networkInterfaces = systemInfoLoader.getHardware().getNetworkIFs();
@@ -86,10 +86,10 @@ public abstract class InformationLoader {
 		systemInformation.physicalMemoryUsed.addValue(systemInformation.physicalMemoryTotal - systemInfoLoader.getHardware().getMemory().getAvailable());
 		double[] loadPerCore = systemInfoLoader.getHardware().getProcessor().getProcessorCpuLoadBetweenTicks();
 		for (int i = 0; i < loadPerCore.length; i++) {
-			systemInformation.cpuUsagePerCore[i].addValue(Math.round(loadPerCore[i] * Config.DOUBLE_TO_LONG));
+			systemInformation.cpuUsagePerCore[i].addValue((short) Math.round(loadPerCore[i] * Config.DOUBLE_TO_LONG));
 		}
 		systemInformation.cpuUsageTotal.addValue(
-				Math.round(systemInfoLoader.getHardware().getProcessor().getSystemCpuLoad() * Config.DOUBLE_TO_LONG));
+				(short) Math.round(systemInfoLoader.getHardware().getProcessor().getSystemCpuLoad() * Config.DOUBLE_TO_LONG));
 
 		final int deadKeepTime = Config.getInt(Config.KEY_DEAD_PROCESS_KEEP_TIME) * 1000;
 		systemInformation.deadProcesses.removeIf(process -> System.currentTimeMillis() - process.deathTimestamp > deadKeepTime);
